[ refactor ]  server remote data schema 분리
