+++ src/types/remote/request.ts
@@ -1,0 +1,13 @@
+export interface PatchingUserPassword {
+  email: string;
+  newPassword: string;
+}
+
+export interface PatchingUserNickName {
+  nickname: string;
+}
+
+export interface PostingVote {
+  ballotTopicId: string;
+  ballotItemId: string;
+}
